requirements.txt
apache-airflow
apache-airflow-providers-google
google-cloud-storage
google-cloud-bigquery
pandas
numpy
pyspark==3.4.3
apache-airflow-providers-apache-spark
